requirements.txt
python-telegram-bot>=20.0
crewai
openai
sqlite-utils
python-dotenv 
singleton_decorator 
google-genai 
neo4j
langchain
langchain-community
langchain-core 